[AMDGPU] Keep profiling data quality out of runtime status
Some ROCm queue profiler configurations can produce zero or inverted HSA completion-signal dispatch timestamps. The runtime cannot cheaply query that up front, and applications should not have to treat profile flush/end as conditionally unsupported based on debug payload quality.

Leave the AMDGPU runtime profile drain as a straight copy of completed dispatch records. Metadata and filter tests validate that records, relationships, and clock-correlation metadata are emitted without making HSA dispatch ticks a universal correctness requirement. Timing-specific tooling can classify unusable timestamp payloads while already walking the decoded records.

Split AMDGPU command-buffer profiling tests out of host_queue_command_buffer_test into a profiling-specific target and local test utility headers. The core command-buffer test no longer carries the profile sink, counter selection, or profiling failure-injection scaffolding.

Keep the ThreadSanitizer exclusion tied to the profiling-specific target after the split. This preserves ASAN/UBSAN coverage while acknowledging that TSAN still cannot model the host queue notification-ring pressure path.

Keep the Linux GPU workflow's rocminfo report diagnostic. Packaged rocminfo can fail before CTest on a runner even though the installed GPU tests are the actual gate for target usability, so report it as a warning and let the tests decide.
